refactor(freertos10): move the dispatch of an interrupt to the port
The table of handlers, the two dispatchers and the registration go to
os/freertos10, and the platform keeps the entry of its vector table and the
enabling of the line in its controller. The depth of a stack is a size_t, the
type FreeRTOS takes when uint16_t is too restrictive.

+++ os/freertos10/src/task.c
@@ -54,7 +54,7 @@ void termina__os__task__init(const termina__id_t task_id,
 
     UBaseType_t task_priority = termina__os__task__priority2freertos(task->priority);
 
-    uint16_t stack_size_in_words = task->stack_size / sizeof(StackType_t);
+    size_t stack_size_in_words = task->stack_size / sizeof(StackType_t);
 
     // Consider the case where the stack size is not multiple of StackType_t
     if (task->stack_size % sizeof(StackType_t) != 0) {
